Conflicts:
	horde/lib/Application.php
	horde/package.xml
	imp/docs/CHANGES
	imp/lib/Application.php
	imp/package.xml
	ingo/lib/Application.php
	ingo/package.xml
	kronolith/lib/Application.php
	kronolith/package.xml
	mnemo/lib/Application.php
	mnemo/package.xml
	nag/lib/Application.php
	nag/package.xml
	turba/lib/Application.php
	turba/package.xml

+++ whups/lib/Driver/Sql.php
@@ -562,12 +562,14 @@ class Whups_Driver_Sql extends Whups_Driver
             }
         }
 
-        try {
-            $sql = 'DELETE FROM whups_transactions WHERE transaction_id IN '
-                . '(' . str_repeat('?,', count($txs) - 1) . '?)';
-            $this->_db->delete($sql, $txs);
-        } catch (Horde_Db_Exception $e) {
-            throw new Whups_Exception($e);
+        if (!empty($txs)) {
+            try {
+                $sql = 'DELETE FROM whups_transactions WHERE transaction_id IN '
+                    . '(' . str_repeat('?,', count($txs) - 1) . '?)';
+                $this->_db->delete($sql, $txs);
+            } catch (Horde_Db_Exception $e) {
+                throw new Whups_Exception($e);
+            }
         }
     }
